+++ packages/expo-router/src/react-navigation/bottom-tabs/utils/useBottomTabBarHeight.tsx
@@ -1,10 +1,10 @@
 'use client';
-import * as React from 'react';
+import { use } from 'react';
 
 import { BottomTabBarHeightContext } from './BottomTabBarHeightContext';
 
 export function useBottomTabBarHeight() {
-  const height = React.useContext(BottomTabBarHeightContext);
+  const height = use(BottomTabBarHeightContext);
 
   if (height === undefined) {
     throw new Error(