+++ packages/expo-router/src/react-navigation/bottom-tabs/views/BottomTabBar.tsx
@@ -1,5 +1,5 @@
 'use client';
-import React from 'react';
+import React, { use } from 'react';
 import {
   Animated,
   type LayoutChangeEvent,
@@ -171,7 +171,7 @@ export function BottomTabBar({ state, navigation, descriptors, insets, style }: 
 
   const isKeyboardShown = useIsKeyboardShown();
 
-  const onHeightChange = React.useContext(BottomTabBarHeightCallbackContext);
+  const onHeightChange = use(BottomTabBarHeightCallbackContext);
 
   const shouldShowTabBar = !(tabBarHideOnKeyboard && isKeyboardShown);