+++ packages/expo-router/src/react-navigation/core/useNavigationHelpers.tsx
@@ -1,5 +1,6 @@
 'use client';
 import * as React from 'react';
+import { use } from 'react';
 
 import {
   CommonActions,
@@ -44,7 +45,7 @@ export function useNavigationHelpers<
   router,
   stateRef,
 }: Options<State, Action>) {
-  const parentNavigationHelpers = React.useContext(NavigationContext);
+  const parentNavigationHelpers = use(NavigationContext);
 
   return React.useMemo(() => {
     const dispatch = (op: Action | ((state: State) => Action)) => {