+++ packages/expo-router/src/react-navigation/native/useLocale.tsx
@@ -1,5 +1,5 @@
 'use client';
-import * as React from 'react';
+import { use } from 'react';
 
 import { LocaleDirContext } from './LocaleDirContext';
 
@@ -7,7 +7,7 @@ import { LocaleDirContext } from './LocaleDirContext';
  * Hook to access the text direction specified in the `NavigationContainer`.
  */
 export function useLocale() {
-  const direction = React.useContext(LocaleDirContext);
+  const direction = use(LocaleDirContext);
 
   if (direction === undefined) {
     throw new Error(