+++ packages/expo-router/src/react-navigation/native/useScrollToTop.tsx
@@ -1,5 +1,6 @@
 'use client';
 import * as React from 'react';
+import { use } from 'react';
 import type { ScrollView } from 'react-native';
 
 import {
@@ -53,7 +54,7 @@ function getScrollableNode(ref: React.RefObject<ScrollableWrapper>) {
 }
 
 export function useScrollToTop(ref: React.RefObject<ScrollableWrapper>) {
-  const navigation = React.useContext(NavigationContext);
+  const navigation = use(NavigationContext);
   const route = useRoute();
 
   if (navigation === undefined) {