+++ packages/expo-router/src/react-navigation/stack/utils/useCardAnimation.tsx
@@ -1,10 +1,10 @@
 'use client';
-import * as React from 'react';
+import { use } from 'react';
 
 import { CardAnimationContext } from './CardAnimationContext';
 
 export function useCardAnimation() {
-  const animation = React.useContext(CardAnimationContext);
+  const animation = use(CardAnimationContext);
 
   if (animation === undefined) {
     throw new Error("Couldn't find values for card animation. Are you inside a screen in Stack?");